Add files via upload
Here is the file with everyone's functions added into it.  I changed only a couple things, like removing the references &.
Checked and compiled this file, should work if everyone wants to test it.

We are only not passing test 13, if anyone wants to look at it and try to figure out where the error is.  I haven't found it.